implementation of seperation widget of author_view
