+++ lib/features/ets/events/author/author_viewmodel.dart
@@ -1,11 +1,8 @@
-// Package imports:
 import 'package:ets_api_clients/models.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 import 'package:infinite_scroll_pagination/infinite_scroll_pagination.dart';
 import 'package:stacked/stacked.dart';
-
-// Project imports:
 import 'package:notredame/features/app/repository/author_repository.dart';
 import 'package:notredame/features/app/repository/news_repository.dart';
 import 'package:notredame/utils/locator.dart';
@@ -14,19 +11,13 @@ class AuthorViewModel extends BaseViewModel implements Initialisable {
   final AuthorRepository _authorRepository = locator<AuthorRepository>();
   final NewsRepository _newsRepository = locator<NewsRepository>();
 
-  /// Localization class of the application.
   final AppIntl appIntl;
-
   final String authorId;
 
-  /// Author
   Organizer? _author;
-
-  /// Return the author
   Organizer? get author => _author;
 
-  final PagingController<int, News> pagingController =
-      PagingController(firstPageKey: 1);
+  final PagingController<int, News> pagingController = PagingController(firstPageKey: 1);
 
   AuthorViewModel({required this.authorId, required this.appIntl});
 
@@ -35,7 +26,6 @@ class AuthorViewModel extends BaseViewModel implements Initialisable {
 
   @override
   Future<void> initialise() async {
-    // This will be called when init state cycle runs
     pagingController.addPageRequestListener((pageKey) {
       fetchPage(pageKey);
     });
@@ -43,8 +33,7 @@ class AuthorViewModel extends BaseViewModel implements Initialisable {
 
   Future<void> fetchPage(int pageNumber) async {
     try {
-      final pagination = await _newsRepository.getNews(
-          pageNumber: pageNumber, organizerId: authorId);
+      final pagination = await _newsRepository.getNews(pageNumber: pageNumber, organizerId: authorId);
       final isLastPage = pagination?.totalPages == pageNumber;
       if (isLastPage) {
         pagingController.appendLastPage(pagination?.news ?? []);
@@ -58,16 +47,17 @@ class AuthorViewModel extends BaseViewModel implements Initialisable {
   }
 
   void notifyMe() {
-    // TODO activate/deactivate notifications
     isNotified = !isNotified;
     if (isNotified) {
       Fluttertoast.showToast(
-          msg: appIntl.news_author_notified_for(author?.organization ?? ""),
-          toastLength: Toast.LENGTH_LONG);
+        msg: appIntl.news_author_notified_for(author?.organization ?? ""),
+        toastLength: Toast.LENGTH_LONG,
+      );
     } else {
       Fluttertoast.showToast(
-          msg: appIntl.news_author_not_notified_for(author?.organization ?? ""),
-          toastLength: Toast.LENGTH_LONG);
+        msg: appIntl.news_author_not_notified_for(author?.organization ?? ""),
+        toastLength: Toast.LENGTH_LONG,
+      );
     }
   }
