+++ lib/features/ets/events/author/widget/back_button_widget.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class BackButtonWidget extends StatelessWidget {
+  const BackButtonWidget({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return IconButton(
+      icon: const Icon(Icons.arrow_back),
+      onPressed: () => Navigator.of(context).pop(),
+    );
+  }
+}
